requirements.txt
Django==1.10.1
django-ajax-selects==1.5.0